Only copy stripy json if exists

+++ cpg_workflows/jobs/stripy.py
@@ -101,7 +101,13 @@ def stripy(
     ls $BATCH_TMPDIR/
 
     cp $BATCH_TMPDIR/{sequencing_group.id}__{sequencing_group.external_id}.cram.html {j.out_path}
-    cp $BATCH_TMPDIR/{sequencing_group.id}__{sequencing_group.external_id}.cram.json {j.json_path}
+
+    # only copy the json if it exists
+    if [ -f $BATCH_TMPDIR/{sequencing_group.id}__{sequencing_group.external_id}.cram.json ]; then
+        cp $BATCH_TMPDIR/{sequencing_group.id}__{sequencing_group.external_id}.cram.json {j.json_path}
+    else
+        touch {j.json_path}
+    fi
 
     """
